Register the default adapter only for Archetypes base content.
We registered it for Interface, which meant 'IOptionalHighlight(some object)' always gave a result.
That is not wanted.  For example for the Plone Site as object, do_highlight would then call getField, which would always fail, and fallback to the default.
Now we cut that short, which should give a tiny performance improvement.
This is all in preparation for adding a dexterity behavior.

Test coverage is at 100 percent now!